Windows changes.

+++ src/iot/uatypes/Node.cpp
@@ -80,7 +80,7 @@ namespace Jde::Iot{
 	}
 	NodeId::NodeId( Proto::ExpandedNodeId&& x )ι{
 		const auto& proto = x.node();
-		nodeId.namespaceIndex = proto.namespace_index();
+		nodeId.namespaceIndex = (int16)proto.namespace_index();
 		if( proto.has_numeric() ){
 			nodeId.identifierType = UA_NodeIdType::UA_NODEIDTYPE_NUMERIC;
 			nodeId.identifier.numeric = proto.numeric();
